refactored switch component
